Final version of comments and image profile.
