+++ Talento/Controllers/ChangeImageProfileController.cs
@@ -34,7 +34,7 @@ namespace Talento.Controllers
 
                 if (appUser.ImageProfile == null)
                 {
-                    i = Image.FromFile(Server.MapPath("~/Content/Images/iconuseralien.png"));
+                    i = Image.FromFile(Server.MapPath("~/Content/Images/alien1.png"));
                 }
                 else
                 {
