+++ Talento/Views/Candidate/Details.cshtml
@@ -126,7 +126,8 @@
                                                 <h6 class="text-muted time">@x.Date.ToLongDateString() - @x.Date.ToShortTimeString()</h6>
                                             </div>
                                         </div>
-                                        <div class="post-description" style="overflow:auto">
+                                        <div class="post-description">
+                                            <br/>
                                             <span>@x.Content</span>
                                         </div>
                                     </div>
