+++ Talento/Views/Positions/Details.cshtml
@@ -132,7 +132,6 @@
                             <i class="material-icons md-24">delete</i>
                         </a>
                     </td>
-
                 </tr>
             }
         </table>
@@ -450,11 +449,11 @@
                         <div class="col-md-8">
                             <div class="panel panel-white post panel-shadow">
                                 <div class="post-heading">
-                                    <div class="pull-left image">
+                                    <div class="pull-left image" style="width:20%">
                                         <img src="https://cdn0.iconfinder.com/data/icons/alien-space/512/call_center_help_desk_reception_hotline_number_emergency_service_phone_operator_support_chat_alien-512.png"
                                              class="img-circle avatar" alt="user profile image"/>
                                     </div>
-                                    <div class="pull-left meta">
+                                    <div class="pull-left meta" style="width:70%">
                                         <div class="title h5">
                                             <span><b>@x.User.Email</b></span>
                                             @if (x.CandidateId != null)
@@ -469,7 +468,6 @@
                                         <h6 class="text-muted time">@x.Date.ToLongDateString() - @x.Date.ToShortTimeString()</h6>
                                     </div>
                                 </div>
-                                <br/>
                                 <div class="post-description">
                                     <span>@x.Content</span>
                                 </div>
