+++ Talento/Views/Shared/Dashboard/_PartialTopNavigation.cshtml
@@ -15,7 +15,7 @@
                             var imgSrc = "";
                             if ((byte[])ViewData["Image"]==null)
                             {
-                                imgSrc = "/Content/Images/iconuseralien.png";
+                                imgSrc = "/Content/Images/alien1.png";
                             }
                             else
                             {
